Writing comments and moved the users table into a separate file
